advanced_distributed_workers/requirements.txt
flask==2.3.3
pytest==7.4.0
pytest-flask==1.2.0
requests==2.31.0
pytest-cov==4.1.0
python-dotenv==1.0.0
